The problem with missing lines due to the lack of pen thickness has been solved, minor bugs have been fixed

+++ DesktopEditor/raster/Metafile/Common/MetaFileObjects.h
@@ -77,7 +77,7 @@ namespace MetaFile
 
 		virtual int          GetColor() = 0;
 		virtual unsigned int GetStyle() = 0;
-		virtual unsigned int GetWidth() = 0;
+		virtual double GetWidth() = 0;
 	};
 }
